fix gas limit from user input
